Add Batcher Signature to Espresso transaction
In order to be able to verify that a transaction coming from Espresso's finalized
transaction list was originally sourced from the op-batcher, we need a way to
verify that the transaction did actually come from the op-batcher.  To facilitate
this we want the transaction submitted to espresso to be signed by the op-batcher
itself.

The majority of this change is a result of attempting to forward the signing
functions to the place where we need them, while also maintaining no ability to
have direct memory access to any potentially stored private key.

This change signs the payload with the desired private key before it is submitted
to Espresso.

Update comments on `clientSigner` and `privateKeySigner`

The comment on `privateKeySigner` is for different code altogether.  It should
be replaced with a comment that applies to `privateKeySigner` itself.  Additionally
the comments could use some more explicit purpose.

This change updates the comments on `privateKeySigner` and `clientSigner` that
help to      explain what they are intended for.

Fix ChainSigner not being set for Config and DriverSetup

The ChainSigner is created, but it's never assigned to Config and for DriverSetup.
This leads to null pointer calls when trying to sign the payload for the chain.

Clean up comment that had an extra "creates a"