Add Batcher Signature to Espresso transaction
In order to be able to verify that a transaction coming from Espresso's finalized
transaction list was originally sourced from the op-batcher, we need a way to
verify that the transaction did actually come from the op-batcher.  To facilitate
this we want the transaction submitted to espresso to be signed by the op-batcher
itself.

The majority of this change is a result of attempting to forward the signing
functions to the place where we need them, while also maintaining no ability to
have direct memory access to any potentially stored private key.

This change signs the payload with the desired private key before it is submitted
to Espresso.