Deterministic L1 starting block selection (Config in advance) (#342)
Changes the way the migration script works so that it does not require the L2 fork block time as an input. This frees us from trying to predict a safe L2 fork block time that falls after the final Celo L1 block but not so far in the future that we end up with the L2 not producing blocks once started because it is waiting for that time.

Also changes the way we select the l1 starting block, previously we were selecting it to be the latest block of the L1 but this was not safe since that block would not have been finalized and it was also not deterministic, since the value depended on when the migration script was run. Now we select the most recent L1 block that was finalized at the time of the L2 fork block (or if there is none the closest we can find), which is deterministic because it is based on the L2 fork block time that is based on the last Celo L1 block time.

These changes allow us to provide all required config for a migration well in advance of the migration time and also allows partners to perform the migration in parallel with our migration, as opposed to having to wait for our migration to finish before they can start their migration.

A description of ethereum's finalization process can be found here - https://eth2book.info/capella/part2/consensus/casper_ffg/
and here - https://ethos.dev/beacon-chain

The approach we use here is to use the finality_checkpoints to find the most recent finalized slot for a given slot, the finality checkpoints don't change so we can easily historically find out what was the latest finalized slot at any point in the history of the chain.

Fixes https://github.com/celo-org/celo-blockchain-planning/issues/908