Config in advance fix map lookups (#354)
Don't use *big.Int as map key for beacon genesis.

big.Ints are not directly comparable in go since they contain a slice,
meaning they cannot be used as a map key, using a pointer to a big.Int
results in pointer equality being used for map lookups which is not what
we wanted.

Before this change all lookups for genesis start time would have
resulted in 0.