Fix version and configuration mismatch
The rebase from celo-tip made the celo-integration branch
go back a version on`github.com/minio/minio-go/v7` from
`v7.0.85` to `v7.0.84`.  Additionally, the devnet configuration
files ended up dropping the `participants` configurations.

This commit updates the version to what it was, and re-adds
the participants configurations in the devnet files.