Update the run_all_tests.sh file
The run_all_tests.sh file was initially setup from a prior integration
and potentially reflects a project state that no longer exists.  For
example `proxyd` and `indexer` no longer seem to be folders that
exist in the project.  Additionally, there are a few complex commands
to run all of the `test` make file commands. However not all of these
tests are run in the same way. Additionally, these tests are not
necessarily run in isolation from each other.  There are
a few tests that `test-unit` run that require the existence of a
`forge-artifacts` file which can be populated by another test.

This commit attempts to specify all of the tests in an unambiguous
sequential  manner with the correct `Makefile` command, while
keeping the early exit for easier test inspection.