Remove address from Sign method on ChainSigner
The address being passed to the ChainSigner's Sign function is likely the incorrect
address.  Additionally, the address is already possible to be known already and is
therefore unnecessary to pass in.

The change removes the need to pass in the address of the key that is being used to
sign the transactions.

Additionally, tests have been added to verify that signature verification works as expected.