Fix public address derivation in `testChainSignerSign`
The equality check on espresso_test.go:81 was a copy of espresso_test.go:68
as correctly pointed out by @dalinsubjam.  This is not what we want.  Instead we
want to be able to derive the address from the public key that is recovered from
the signature, and to test that against our expected address.

This modifies the logic to manually derive the address from the recovered public
key utilizing publicly available functions within the go-ethereum repository. As
a result, the test has been corrected, and now passes successfully.