Add helper files to configure dev environment with espresso-dev-node
In order to help setup repeatable testing in an environment that is configured and can
be spun up quickly, we need to be able to quickly and easily launch a testing
environment with the configuration we desire.  We can take advantage of the e2e
utils that optimism has provided, and add the ability to spin up our dev-node like
we do for kurtosis.  We are also able to take advantage of the espresso-dev-node
being able to be run within a docker container as a stand-alone service.

This commit adds a basic smoke test for spinning up an espresso-dev-node with
the e2e system.  This should prove as a working proof of concept that additional
tests can utilize to quickly spin up an environment for testing.

This commit also adds a few utility functions, methods, and types for convenience
and to separate concerns in a clean manner.