Fix dev net not submitting to Espresso
The E2E Dev Net was not communicating with Espresso despite initial efforts.
This was due to the UseAltDa configuration options not being set after the
Start method has been called.  Using the debugger it was clear that this was
due to the configuration being incomplete.

This commit fixes these issues.  Additionally, it further separates the launch
concerns into some setup launch options.