Fix cleanup in `espressoBatchLoadingLoop`
The method `espressoBatchLoadingLoop` in the `BatchSubmitter` does not close the
`publishSignal` channel, like the non-espresso logic method `blockLoadingLoop` does
when it exits.  This prevents the Batcher from terminating when it is told to `Stop`.

This change adds a `defer close(publishSignal)` to the start of the
`espressoBatchLoadingLoop` to ensure that the `publishSignal` channel is closed
when it exists.