Remove AltDA usage for Espresso
The AltDA requirement for Espresso is no longer necessary.  It is now considered
separate behavior, and can just cause configuration / confirmations to be confused.
It should be removed from the e2e system local devnet tests so it will not be a factor
in things running correctly.

This commit removes the AltDA usage that was previously required to run the
Espresso Dev Node