Replace the context passed when submitting to Espresso
The context passed to Espresso.SubmitTransaction and
Espresso.FetchTransactionByHash comes from the `BatcherSubmitter`'s `shutdownCtx`
field.  Originally, I was planning on adding another `case` in the `select` statement to
explicitly account for the `shutdownCtx`, as having the `FetchTransactionByHash` would
potentially infinitely loop if `ctx` was not derived from `shutdownCtx`.  However, in
chasing down the context being passed in, I was able to determine that these are already
the same context instances as the `shutdownCtx`.

The passed `ctx` ultimately comes from a call to `espressoBatchQueueingLoop` which
is invoked on `op-batcher/batcher/driver.go:186`.  The passed context is already the
`shutdownCtx` here, so any explicit reference to the `shutdownCtx` within the scope
of the method, makes less sense.

This commit replaces the reference of `l.shutdownCtx` with the passed in `ctx` which
is already the same, or derived from the `l.shutdownCtx`.