Add log parsing utilities for Espresso Dev Node
The light contract address seems to be inconsistent when deploying the Dev Net
with Espresso.  As of now the only reliable way we know of to retrieve the address of
the light client contract, is to retrieve it from the logs of the espresso-dev-node.

This commit adds log parsing utilities to be able to extract deployed contract addresses
from the espresso-dev-node.