Add auto populate light client contract address
The Light Client Contract address has been seen to be inconsistent at times.  After
some discussion with @QuentinI, it seems that the only reliable way to retrieve the
deployed contract addresses is via the logs of the espresso nodes themselves. To
that end, this will read the logs of the `espresso-dev-node` and extract out the
deployed contract addresses before the batcher is deployed, in order to enure
that we have the correct light client address.

Doing this adds a lot of delay in the startup time of the `espresso-dev-node` and
as a result the entire e2e testing framework setup.  So it does not spawn as quickly,
but the retrieved addresses should be correct.