Add support to run the e2e espresso devnet tests for linux
The working solution for spinning up the e2e espresso devnet
actively works on macos platforms, but fails to work on linux
natively.

The reason for this is that we took advantage of a nice feature
that Docker for x (Docker for Mac) has that allows a running
container to connect to the host system via the reserved
DNS entry "host.docker.internal".  This feature works well on
the non-linux docker platforms but is non-existent / not-supported
on linux itself.

If you want to be able to connect / talk to the host machine on
linux, the easiest way is to use the host network mode.  It should
be noted that this network mode doesn't tend to work the non-
native docker platforms, as they host machine tends to refer to the
VM in which they are running.

This change adds support for linux and non-linux by keeping the
existing implementation, but when running on linux, attempts to
run in host mode instead.