Add support for Port remapping for espresso-dev-nodes
The espresso-dev-node on linux runs with the network mode set
to the host.  This prevents us from running multiple instances at
once, as we expect the port that is utilized by the container to be
able to be unique to the running container.  But since the network
mode host shares the networking layer with the host machine, the
same port numbers cannot be utilized at the same time.

In order to get around this, for linux, this change introduces a
port remapping to discover available ports and utilize them instead
of the hard-coded values.  It also adds entries into the port
mapping of the return dev container info so that we maintain the
original port mapping.