Add Espresso Block Conversion Test
In order to better validate that our stored representation of data
from the L2 does not lose information, or result in different
information when stored, it would be beneficial to have a test to
validate these concerns.

This commit adds a test to ensure that our EspressoBlock is able
to be constructed from a `Block` and rollup info  and then
converted back again with the passed rollup info.