Add Deterministic Derivation Caff Node test
We want a test to verify that our Caff Node derivation matches the derivation from the
non-caff node version.  As a starting point, we want to ensure that the initial balance, and
the transaction increments match each step.

This commit adds a helper function to spin up a Caff Node after the e2e system with
espresso dev node has already spun up.

This commit also adds a test to utilize this Caff node and attempts to verify a simple
test of incrementing Alice's balance by 1 Wei.