Revert the increased timeout duration in the waits file
The waits file's timeouts were increased for debugging purposes.  However, it should be
noted that these increased timeouts slow down tests, and if these effects do take too
longer that should be a cause for concern.  If anything, the timeouts should be reduced
further to ensure expedience.

This change reverts the increase of timeouts from 2 minutes to 10 minutes in the
waits.go file.  Thus the timeouts should now be 2 minutes again.