Update  tests to ensure container is stopped
The Docker containers that are spawned by the tests are tied to
the context that are passed in them.  When the context is killed,
the docker container should also die.  However, in practice this
doesn't seem to be the case consistently.

In order to ensure that things are being shutdown consistently
we add a deferred call to the `Stop` and `Close` methods for
both the system and the `espresso-dev-node`.