Add e2e option to allow intercept
In order to effectively test the liveness requirement of our
integration, we want to ensure that we can withstand temporary
outages / downtime and still provide the same finality guarantees.

The ideal way to perform these sort of checks would be to turn
off subsystems within the Espresso environment itself.  However,
this is not currently possible.  The next best thing we can do is
to intercept the request going to Espresso, and handle them
directly or indirectly in order to simulate the desired behavior.

In order to do this we need to be able to intercept the http
requests being made against the Espresso System.

This change allows us to configure an option to intercept
requests going to Espresso's Query Service endpoint, so
that we might decide how to handle them in order to
simulate the desired behavior.