Refactor decision making behavior of intercept
The intercept's interface and handling is a little confusing and unwieldly.
I have made efforts to refactor it so that al ot of the specific details need
not be worried about directly, while still attempting to ensure that the
caller has the ability to override the specific behavior progression he/she
may desire.

The result is more flexibility with less specific configuration required.