Modify the liveness test to speed it up
The liveness test base case runs pretty quickly, but it can be sped up if
we don't run a full L1 -> L2 flow.  If we instead just wait for the receipt on
the L2 Sequencer, then verify that it appears on a L2 Verifier later, we
can end up with a much quicker to run test.

This separates these pieces out for speed