Update liveness test to simulate unreachable service
The acceptance criteria for the test states to only turn the builder off
and on which would only present itself as some submitted transactions
not being returned / confirmed by espresso.  However, the original
criteria also states to turn off random components of the espresso
network, which could present themselves as the services being
unavailable temporarily.

In order to help facilitate this design, a new behavior and handler have
been added to the Intercept to simulate an unreachable service.