Add some unit tests to Espresso Streamer
In order to increase our confidence in our implementation of the
EspressoStreamer it is beneficial to have unit tests that are able to
simulate the expected behavior, and the handling of edge cases we
anticipate (and potentially do not anticipate) occurring.

With this in mind the `EspressoStreamer` has been modified to consume
interface implementations instead of struct instances so these
components are able to be swapped out with others, specifically our
mock implementation so we can more easily and readily control the
state iteration of the system and the interactions between the L1,
L2, and Espresso.