Remove unnecessary for loops in mock source
In review it was pointed out by @QuentinI that incrementing through a for
loop to change the position by one incrementally was unnecessary, and
potentially confusing:
https://github.com/EspressoSystems/optimism-espresso-integration/pull/125#discussion_r2075881991
https://github.com/EspressoSystems/optimism-espresso-integration/pull/125#discussion_r2075743297

The only reason to have an incremental approach like this would be due
some non deterministic reference creation, or if there are side-effects
being triggered by the advancing.  This is not currently the case for
either of these calls, and this makes the for loop entirely unnecessary
and potentially confusing.