Remove SafeL1 from MockStreamerSource
As pointed out by @QuentinI, the reference to the `SafeL1` is not
relevant.  It was originally added as the `Refresh` call made a reference
to it for logging originally.  This reference and log were removed thereby
removing the need for the `SafeL1` tracking at all.

https://github.com/EspressoSystems/optimism-espresso-integration/pull/125#discussion_r2075750502

Replace all former references of the `SafeL1` with teh `FinalizedL1`
within streamer_test.go.