Fix averageDuration calculation
The `averageDuration` calculation is dividing by `N` instead of how many entries we
actually encountered, as pointed out by @dailinsubjam.  While this should be fine if everything
works, it does lead to potential errors in computation.  It **should** be the case that the number
of receipts received should correspond to the number of received submissions, so we should
be able to leverage that number instead.

https://github.com/EspressoSystems/optimism-espresso-integration/pull/126#discussion_r2076812347

Fix the averageDuration calculation.