Revert test to check timing using block hash
When designing the test it was noticed that we weren't able to
verify beyond the first block hash for the transactions submitted.
Instead of using this information as an indication of a failure, primarily
due to noticing that the Caff Node was able to verify the receipts
coming through correctly.  Instead it was assumed that the approach
itself was incorrect.

I ended up going choosing an approach that had some faulty
assumptions.  Namely that the transactions we were submitting were
the only batches / l2 blocks being produced.  This is incorrect as
some observation against the tests indicate that 10-16 blocks are
created initially on launch before we even submit any transactions.

This change reverts the test to utilize the Block Hash, as that does
seem to be a pretty decent check for tracking a transaction's
progress through the system.