Split `Update` method in EspressoStreamer
The `Update` method is doing a lot of things within a single closure
that can make it difficult to track / reason about.  Some of these
concepts can be better isolated to help reduce cognitive load of
reading this implementation, while still persisting the underlying
concepts of what each section is achieving.

This change refactors the `Update` method to move some of these
more isolate concepts together while preserving the original
implementation's behavior.  This is just an organizational change,
and not a functional one.