Update `Update` method to try and ensure batch availability
The `EspressoStreamer` was written with the idea to load blocks
from `HotShot` in discrete batches.  This is a great idea, but it may
break standard consumer assumptions about how streams tend to
work.

In general, it doesn't hold the principal that calling `Update`
will load **AT LEAST** the next available `Batch` into the `Streamer`
or fail due to an unforeseen error or having no actual next `Batch`
be currently available.  I believe this is a reasonable expectation that
`Update` should do it's best to fulfill, as not doing this will just result
in the user calling `Update` and `Next` again to get to this desired
result anyway.

This change refactors `Update` some more, and allows it to process
more Batches to try and ensure that there is a next `Batch` loaded
into the `BatchBuffer`.