Add tests for soft confirmation integrity
Since Espresso provides a different derivation source / authority for the optimism state
it becomes very important to ensure that the soft confirmations that the L2 provides are
not modified based on what Espresso provides.  To put a different way, the information
that Espresso provides should not be able to modify the derivation in any way that would
cause it to deviate from what the L2 Sequencer itself provides.

This commit adds tests to showcase and verify that this integrity is maintained, at a base
level, and again under adversarial conditions.