Add implementation for random signature submission
In the soft confirmation tests we had initially implemented only two of
the proposed three cases for malicious / attack activity for Espresso.
We implement the last activity that will yield a random signature on
the malicious batch in our continued efforts to demonstrate our
reliability.

Additionally, We clean up some code, replace hard coded values with
constants and comments, and adjust the random signature to actually
be random, and not derived from some string.