Adjust VerifyL1OriginFinalizedCheck to reference L1 finalized block
The `TestCaffNodeWaitForFinality` test was a bit flaky as it was utilizing
the L2Verifier's understanding of what the finalized L1 was.  This can be
problematic, and was causing the test to fail occasionally as it can be
behind reality.

To address this, we instead utilize the L1's understanding of what it's
finalized block is.  This is more in line with the spirit of the original
test's intention, and helps the test to pass more consistently.