Refactor local `Update` variables for clarity
Based on feedback provided by @QuentinI, continually utilizing the
variables `start`, and `finish` will provide better clarity over using a
local increment variable `i`:
https://github.com/EspressoSystems/optimism-espresso-integration/pull/129/files#r2089587647

The `i` variables instead of `processRange` has been renamed to be
`height` to better signify it's intention.

The loop inside of `Update` has been reworked to exclusively rely on
`start` and `finish` instead of using a local increment `i` variable.

The `start` position calculated in `computeEspressoBlockHeightsRange`
has been modified to start at the next block, instead of the the
`hotShotPos`, as the `hotShotPos` indicates a block we've already
processed (provided it is not the starting block)