Fix an Espresso Block Processing edge case
There is an edge case where the initial block being processed from
espresso actually contains the next batch.  We need to process the initial
block, so we need to loosen the restrictions on the condition:
`start` >= `finish`.  However, when we loosen this restriction we end up
with the potential to get stuck in an infinite loop if the only block available
from Espresso is its initial block.

We add some safe guards to prevent this edge case while still ensuring
that it processes the initial block at least once.