Add timeouts to soft confirmation tests
The soft confirmation tests can run indefinitely if not limited to an
execution time.  If running well, they should be able to complete and
pass within a reasonable amount of time.

This adds a 5 minute timeout, and handling for a completed context to
the tests to ensure that the test doesn't run indefinitely.