Adjust caff node test for liveness
The liveness criteria of the Caff node has an additional assumption in that
the L2 is only referencing finalized L1 blocks when it builds blocks.  This
assumption did not currently exist within the test.  This has been modified
so that the L2 blocks do reference the L1 Blocks specifically.

Additionally, there is a criteria that we should be able to received finalized
confirmations within 10 seconds.  But based on conversations with
@philippecamacho, anything within the 10 second range should be
sufficient.  So the test has been modified to have a requirement of being
strictly less than 11 seconds.

Update documentation to reflect 11 seconds

Based on the conversion from @philippecamacho the documentation still
references 10 seconds instead of the modified 11 seconds: https://github.com/EspressoSystems/optimism-espresso-integration/pull/142#discussion_r2098677816.

Additionally, a conditional in the test is still checking against the 10
seconds threshold instead of the 11 second one.

There was also a linting error introduced as we aren't inspecting or
referencing a returned `err` value.