Cleanup Benchmarker code
The benchmarker defines quite a lot of things that are not all grouped very
well.  These concepts should be split up into sensible groups.

Add Comments

The comments on the Benchmarker functions and primitives are lacking,
and incomplete.  Comments should be added to annotate, and document
the specific implementation details, and their reason for being.

Modify test benchmark statistics and variation targets

The benchmark statistics being checked against in the test have caff
receipt durations, and l2 verifier receipt durations that are weighted by
the total number of transactions in their respective blocks.  Additionally
these statistics lack any ability to track a block that does not contain the
submitted transactions.  This has been addressed by changing the
statistical measure computation.

The variation targets for the StdDev check were just set to 10% of the
mean.  This is a somewhat difficult target to pin down.  For now they have
been specified to not exceed 2 seconds instead.