Add 100ms delay in repeat transaction verification
When a transaction fails to verify, it will be added into the job queue and
reattempted without any delay between these attempts.  This can end up
with a potential DOS vector for multiple requests being attempted back
to back as soon as they are able to be performed.  There should be a
delay between these attempts, ideally without impacting other requests
that don't need to wait from continuing.

Add a 100ms delay to submission verification re-attempts.  This is not as
optimal as it could be and will prevent other jobs from coming in at the
moment.

Modify Espresso Submitter to respect shutdown

The Espresso Submitter should stop processing things when it is told
to do so by the Batcher.  The shutdownCtx governs whether things should
run or not.

The stop worker path has been modified to just utilize the same
WaitGroup and context as the other processes.