Fix inconsistent transaction count received
We don't have a way of knowing when we're really "done" in terms of
processing batches from the Espresso Streamer.  This means that we
also don't know when it's safe to start checking our invariants.

In order to address this effectively, this change modifies where the
initial signed transactions are created, so we can match the last
transaction in the decoded block.  This way we'll know we're done,
and it can provide a sufficient signal to start the test invariants.