Adjust the success threshold to 17 seconds
The success threshold had been previous bumped up to 11 seconds.
However, even with this additional second, we are still seeing failures.
We still see failures occurring occasionally all the way up to 15 seconds.
In a previously run benchmark test, we saw that the max time for getting
a response when getting from L2 Sequencer Receipt to Caff Derived was
just shy of 17 seconds.  As a result, we are adjusting the acceptance
criteria of this to 17 seconds, while we try to dig further into the issue to
discover where the delays are coming from.