Make Test 2 Pass (#150)
* Fix inconsistent transaction count received

We don't have a way of knowing when we're really "done" in terms of
processing batches from the Espresso Streamer.  This means that we
also don't know when it's safe to start checking our invariants.

In order to address this effectively, this change modifies where the
initial signed transactions are created, so we can match the last
transaction in the decoded block.  This way we'll know we're done,
and it can provide a sufficient signal to start the test invariants.

* Adjust the success threshold to 17 seconds

The success threshold had been previous bumped up to 11 seconds.
However, even with this additional second, we are still seeing failures.
We still see failures occurring occasionally all the way up to 15 seconds.
In a previously run benchmark test, we saw that the max time for getting
a response when getting from L2 Sequencer Receipt to Caff Derived was
just shy of 17 seconds.  As a result, we are adjusting the acceptance
criteria of this to 17 seconds, while we try to dig further into the issue to
discover where the delays are coming from.