Change strategy for Transaction Submission to Espresso (#138)
* Add worker queue transaction submission to Espresso

The Workflow for submitting transactions to Espresso currently will
spawn an unbounded number of goroutines to submit transactions to
Espresso.  This can be incredibly resource heavy, and may end up
actually harming performance due to schedule thrashing.

This change adds a basic worker based implementation for processing
individual job requests for transaction submission and receipt verification.
This should help limit the total number of requests that are being
submitted to Espresso at any given time.

There are improvements that are able to be made to this, but this serves
as a basic implementation that matches what was happening originally
with one delay missing.

* Add 100ms delay in repeat transaction verification

When a transaction fails to verify, it will be added into the job queue and
reattempted without any delay between these attempts.  This can end up
with a potential DOS vector for multiple requests being attempted back
to back as soon as they are able to be performed.  There should be a
delay between these attempts, ideally without impacting other requests
that don't need to wait from continuing.

Add a 100ms delay to submission verification re-attempts.  This is not as
optimal as it could be and will prevent other jobs from coming in at the
moment.

Modify Espresso Submitter to respect shutdown

The Espresso Submitter should stop processing things when it is told
to do so by the Batcher.  The shutdownCtx governs whether things should
run or not.

The stop worker path has been modified to just utilize the same
WaitGroup and context as the other processes.

* Update comments and code based on feedback

The feedback from @philippecamacho has pointed out some
implementation details, and comments that were lacking, unfinished, or
lacked sufficient clarity.  These issues have been addressed where
immediately obvious, and not requiring further clarification.