Specify a `create2Salt` value for espresso deployment
When the `op-deployer` performs the `apply` operation, it utilizes the values in the
`state.json` file to inform how it should behave.  Namely among these is the
`create2Salt` value.  The `op-deployer` utilizes the `create2` method for
performing deployments according to specifications.  This salt value is utilized and
helps to assist in determining the resulting contract addresses.  For convenience, it
would be nice to have some deterministic values to test against for repeatability.

This change modifies the `state.json` file before running `op-deployer` `apply`
in order to ensure that we can see deterministic contract address generation.