Fix op-proposer configuration for docker-compose.yml (#207)
* Update GAME_FACTORY to use proxy dispute game factory address

* Fix incorrect path string in GAME_FACTORY settings

* Add Espresso contract configs to prepare-allocs

The prepare-allocs.sh script doesn't currently have the Espresso Smart Contracts
enabled.

This change enables the contract and sets the preApprovedBatcherKey to work in
a non-TEE environment.

* Modify game-type for op-proposer

The default game type when the `game-type` parameter is unspecified is `0`.  By
default there does not seem to be a `Dispute Game` deployed for `game-type` `0`,
however, other game types do exist, including `game-type` `1`.

This change sets the `game-type` of the `op-proposer` to `1` for the
`docker-compose.yml` file.

* Specify a `create2Salt` value for espresso deployment

When the `op-deployer` performs the `apply` operation, it utilizes the values in the
`state.json` file to inform how it should behave.  Namely among these is the
`create2Salt` value.  The `op-deployer` utilizes the `create2` method for
performing deployments according to specifications.  This salt value is utilized and
helps to assist in determining the resulting contract addresses.  For convenience, it
would be nice to have some deterministic values to test against for repeatability.

This change modifies the `state.json` file before running `op-deployer` `apply`
in order to ensure that we can see deterministic contract address generation.