IA1.2.11 Write scripts and documentation for demo (#211)
* Add devnet test for batcher restart

* Check error returns

* Separate op-geth instances for each L2 node

* Build devnet dockers in CI

* Build op-deployer in CI

* Try larger runner

* Increase test outage and recovery time

* Try to speed up transaction verification

* Do not drop batches before we have seen a finalized L1 block

* Remove unnecessary sleep

* Add scripts for demo and documentation

* Build containers in dependency order

* Don't copy config file into Docker image at build time

* Fix syntax

* Checkout submodules in CI

* Don't copy config file into Docker image at build time

* Remove another COPY

* Run devnet test in separate workflow

* Add exposed ports for more nodes in docker-compose

The `verifier`, `caff-node`, and `sequencer` all do not have their ports forwarded to
the host machine.  Having these ports exposed can make testing / debugging easier
by utilizing the optimism RPC API.

This change exposes these ports.

* Fix CAFF environment variables

The CAFF environment variables utilized in the `docker-compose.yml` need to have
a prefix of `OP_NODE_`, otherwise they will not apply.  These not applying causes
the `caff-node` to never make progress.

This change corrects the misconfigured `caff-node` environment variables, and
cleans up an unnecessary one.

* Set `caff.node` to `true` for `caff-node`

The `caff-node` is meant to be running a derivation based on information retrieved
from Espresso's network.  However, it needs to be enabled to do so.  This change
enables the `caff-node`'s mode by explicitly setting `caff.node` to `true` in the
launch configuration.

* Support alias and input check

* Add scripts and instructions for running utility script

In order to showcase the speed and progress of the Espresso solution with the Caff
node, versus the existing sequencer, or the L1 derived verifier, a script has been
added that utilities `watch` and `tmux` to provide a nice visual comparison between
the three nodes.

* Update scripts after separating geth

* Update readme

* tmux in flake.nix

* Update shell being targeted in get_sync_status.sh

The `get_sync_status.sh` script attempts to target `zsh` residing within `/bin/zsh`,
but this isn't a universally available. It is better to target a more commonly available
like `/bin/bash`.

This change modifies the script to target `/bin/bash` instead of `/bin/zsh`.

---------

Co-authored-by: Jeb Bearer <jeb@espressosys.com>
Co-authored-by: Theodore Schnepper <tschnepper@espressosys.com>
Co-authored-by: Philippe Camacho <philippe@espressosys.com>