Fix inclusiveBlockRange invariant
It is possible for the returned `inclusiveBlockRange` to can return a `start` value that
is greater than the `end` value.  This can be addressed by adding an extra check to
the return statement that causes it.

This changes adds a check to ensure that it doesn't return a start value that actually
follows the end value of the range.