Add a Buffered Streamer around Espresso Streamer for batcher
Because the `EspressoStreamer` is getting `Reset` during the Batcher process when
building a batch to submit to the `L1`, it hinders progress of the chain in a reasonable
amount of time, which ultimately causes it to stop creating non-empty blocks.

There are a number of factors that are contributing to this issue, but ultimately the
`Reset` is causing the `EspressoStreamer` to restart from `0` and it takes too long
before it catches back up to the next expected batch.

To remedy this, a Buffer can be used to mitigate this `Reset` and revert to a point
that is much closer to the desired next batch.

In testing it has been observed that the `SafeL2` can sometimes move backwards.  To
safe guard against this, it is better to `Reset` to the `FinalizedL2` position instead of
the `SafeL2` as this behavior has not been observed there.