Rename EspressoStreamer and EspressoStreamerIFace
Based on feedback provided by @QuentinI in PR review:
https://github.com/EspressoSystems/optimism-espresso-integration/pull/230#discussion_r2345028112

The name `EspressoStreamerIFace` is quite a long name, and the `IFace` suffix isn't
necessary since one could tell it's an `interface` by inspection, or using an `LSP`. The
feedback provided by @QuentinI suggested to rename `EspressoStreamerIFace` to
just `EspressoStreamer` so that it falls in line with our other code approaches.

This change renames `EspressoStreamer` to `BatchStreamer`

This change renames `EspressoStreamerIFace` to `EspressoStreamer`.